Update setup.py
Bring xhtml2pdf to the current 0.0.6 version. 

It is compatible with the project and it solves issues when generating reports with large amount of rows that caused "Division by zero" errors caused by the CSS even though the css was valid. 

See for reference on the error:
https://github.com/chrisglass/xhtml2pdf/issues/20

I have had to revisit my css as the structure of the generation is slightly different, but not much work required.